+++ src/php/Game/Rules/RuleSetInterface.php
@@ -2,9 +2,12 @@
 
 namespace Hexammon\HexoNards\Game\Rules;
 
+use Hexammon\HexoNards\Game\Move\MoveGeneratorInterface;
+
 interface RuleSetInterface
 {
 
     public function getInitialSetting(): InitialSettingInterface;
 
+    public function getMoveGenerator(): MoveGeneratorInterface;
 }