+++ src/php/Game/Rules/RuleSetInterface.php
@@ -2,6 +2,7 @@
 
 namespace Hexammon\HexoNards\Game\Rules;
 
+use Hexammon\HexoNards\Game\Game;
 use Hexammon\HexoNards\Game\Move\MoveGeneratorInterface;
 
 interface RuleSetInterface
@@ -10,4 +11,6 @@ interface RuleSetInterface
     public function getInitialSetting(): InitialSettingInterface;
 
     public function getMoveGenerator(): MoveGeneratorInterface;
+
+    public function isGameOver(Game $game): bool;
 }