+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
  
+## [Unreleased] 
+### Fixed
+- One unit army moving: bug with zero-size army after move.
+
 ## 1.0.0-alpha.4 - 2017-11-29
 ### Added 
 - GameOverDetectorInterface with Classic implementation.