+++ CHANGELOG.md
@@ -9,7 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - One unit army moving: bug with zero-size army after move.
 - Check number of units in MoveArmy action.
 - Remained after moving army bug.
-- Check that tile has army on move action.   
+- Check that tile has army on move action.
+
+### Changed
+- Game accept RuleSetInterface instance in third argument instead MoveGenerator. 
+
+### Added
+- RuleSetInterface - aggregator of rules package classes.     
 
 ## 1.0.0-alpha.4 - 2017-11-29
 ### Added